requirements.txt
asttokens==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
fonttools==4.58.5
imageio==2.37.0
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
lazy_loader==0.4
matplotlib==3.10.3
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.1
packaging==25.0
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pywin32==310
pyzmq==27.0.0
scikit-image==0.25.2
scipy==1.16.0
six==1.17.0
stack-data==0.6.3
tifffile==2025.6.11
tornado==6.5.1
traitlets==5.14.3
wcwidth==0.2.13
widgetsnbextension==4.0.14
